// src/api/types.generated.ts
export interface components {
    schemas: {
        UserResponseDto: {
            /** @example 9f6c2fc9-936e-4dd9-9610-fd6f7da9425a */
            id: string;
            /** @example user@example.com */
            email?: string;
            /** @example Андрій Юрчук */
            name?: string;
            /** @example https://lh3.googleusercontent.com/a/ACg8ocIpKpeLRN4eCjevhdvABBECYo_UOmRvEqS-chtHoYw83ww9dQ=s96-c */
            avatarUrl?: string;
        };
// ...
        GoogleSigninResponseDto: Record<string, never>;
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}